Initial implementation of tech tags
Content creators can give any outfit, weapon or ship "tech: " tags,
i.e., any attribute that begins with the string "tech: ". A number isn't
required, but can be optionally added. All tech tags you have on
outfits/ships in your active fleet are kept up-to-date in conditions, so
can be used for mission logic. e.g. Quarg outfits and ships could be
given "tech: quarg" and missions could trigger off having any of them.

Note: initial implementation is a proof of concept. Seems to work, and
I've thought through the solution, but it might not be the best code.