Create an auto condition for your credit balance
This is parallel to the "net worth" auto condition, and allows missions
to require you have >=, <, etc. a certain credit balance. This is
necessary for negative payment missions (missions where you pay a bribe,
etc.) to be handled correctly if you can't afford it. Without this, they
just don't appear. With this, you can create a mirror 'blocking' mission
if you don't have enough money.